EmberIntlIndex: Merge back into EmberIntlIndexExtension class
